Convert overview page to laravel
