+++ html/mix-manifest.json
@@ -3,10 +3,10 @@
     "/css/app.css": "/css/app.css?id=5da3bf931f2f95a17884",
     "/js/manifest.js": "/js/manifest.js?id=3c768977c2574a34506e",
     "/js/vendor.js": "/js/vendor.js?id=29212a758157c575d7f8",
-    "/js/lang/de.js": "/js/lang/de.js?id=18b0b0e06813d1afed92",
-    "/js/lang/en.js": "/js/lang/en.js?id=41955f4356f6511b76b6",
-    "/js/lang/fr.js": "/js/lang/fr.js?id=4f329163511445d92a17",
-    "/js/lang/ru.js": "/js/lang/ru.js?id=e10e85f321f1395378b6",
-    "/js/lang/uk.js": "/js/lang/uk.js?id=c8d4937e3ca47b60b7ac",
-    "/js/lang/zh-TW.js": "/js/lang/zh-TW.js?id=8020327e80ac10c0c615"
+    "/js/lang/de.js": "/js/lang/de.js?id=8c1c390a5f45ac01d54b",
+    "/js/lang/en.js": "/js/lang/en.js?id=14d93644610d1daa14af",
+    "/js/lang/fr.js": "/js/lang/fr.js?id=21b66f3f48a3f67ba443",
+    "/js/lang/ru.js": "/js/lang/ru.js?id=1a4435677a27aa615af1",
+    "/js/lang/uk.js": "/js/lang/uk.js?id=6ce358a9e9387947ee03",
+    "/js/lang/zh-TW.js": "/js/lang/zh-TW.js?id=3ad4a3e6a5a59165e806"
 }
